refactor: use AbstractDatabaseStrategy tests: refactor to use common initialization
