+++ src/Table/Strategy/Exasol.php
@@ -2,17 +2,10 @@
 
 namespace Keboola\InputMapping\Table\Strategy;
 
-use Keboola\InputMapping\Table\Options\InputTableOptions;
-
-class Exasol extends Snowflake
+class Exasol extends AbstractDatabaseStrategy
 {
-    public function downloadTable(InputTableOptions $table)
+    protected function getWorkspaceType()
     {
-        $loadOptions = $table->getStorageApiLoadOptions($this->tablesState);
-        $this->logger->info(sprintf('Table "%s" will be copied.', $table->getSource()));
-        return [
-            'table' => [$table, $loadOptions],
-            'type' => 'copy',
-        ];
+        return 'exasol';
     }
 }
