+++ src/Table/Strategy/Redshift.php
@@ -2,26 +2,10 @@
 
 namespace Keboola\InputMapping\Table\Strategy;
 
-use Keboola\InputMapping\Helper\LoadTypeDecider;
-use Keboola\InputMapping\Table\Options\InputTableOptions;
-
-class Redshift extends Snowflake
+class Redshift extends AbstractDatabaseStrategy
 {
-    public function downloadTable(InputTableOptions $table)
+    protected function getWorkspaceType()
     {
-        $tableInfo = $this->clientWrapper->getBasicClient()->getTable($table->getSource());
-        $loadOptions = $table->getStorageApiLoadOptions($this->tablesState);
-        if (LoadTypeDecider::canClone($tableInfo, 'redshift', $loadOptions)) {
-            $this->logger->info(sprintf('Table "%s" will be cloned.', $table->getSource()));
-            return [
-                'table' => $table,
-                'type' => 'clone'
-            ];
-        }
-        $this->logger->info(sprintf('Table "%s" will be copied.', $table->getSource()));
-        return [
-            'table' => [$table, $loadOptions],
-            'type' => 'copy',
-        ];
+        return 'redshift';
     }
 }
